Created QChatCore that has the common functionality of Server and Client.  Fixed communication time by sleeping between polls of the connection
